Split up action menu in dataset page into multiple buttons.
